Split test_fields file to some files
