+++ tests/fields/base.py
@@ -1,0 +1,41 @@
+import pytest
+
+
+from rest_framework import serializers
+
+
+# Tests for field input and output values.
+# ----------------------------------------
+def get_items(mapping_or_list_of_two_tuples):
+    # Tests accept either lists of two tuples, or dictionaries.
+    if isinstance(mapping_or_list_of_two_tuples, dict):
+        # {value: expected}
+        return mapping_or_list_of_two_tuples.items()
+    # [(value, expected), ...]
+    return mapping_or_list_of_two_tuples
+
+
+class FieldValues:
+    """
+    Base class for testing valid and invalid input values.
+    """
+
+    def test_valid_inputs(self):
+        """
+        Ensure that valid values return the expected validated data.
+        """
+        for input_value, expected_output in get_items(self.valid_inputs):
+            assert self.field.run_validation(input_value) == expected_output
+
+    def test_invalid_inputs(self):
+        """
+        Ensure that invalid values raise the expected validation error.
+        """
+        for input_value, expected_failure in get_items(self.invalid_inputs):
+            with pytest.raises(serializers.ValidationError) as exc_info:
+                self.field.run_validation(input_value)
+            assert exc_info.value.detail == expected_failure
+
+    def test_outputs(self):
+        for output_value, expected_output in get_items(self.outputs):
+            assert self.field.to_representation(output_value) == expected_output
