+++ tests/fields/test_date_field.py
@@ -1,0 +1,68 @@
+import datetime
+
+from django.utils import six
+
+from rest_framework import serializers
+
+from .base import FieldValues
+
+
+class TestDateField(FieldValues):
+    """
+    Valid and invalid values for `DateField`.
+    """
+    valid_inputs = {
+        '2001-01-01': datetime.date(2001, 1, 1),
+        datetime.date(2001, 1, 1): datetime.date(2001, 1, 1),
+    }
+    invalid_inputs = {
+        'abc': ['Date has wrong format. Use one of these formats instead: YYYY[-MM[-DD]].'],
+        '2001-99-99': ['Date has wrong format. Use one of these formats instead: YYYY[-MM[-DD]].'],
+        datetime.datetime(2001, 1, 1, 12, 00): ['Expected a date but got a datetime.'],
+    }
+    outputs = {
+        datetime.date(2001, 1, 1): '2001-01-01',
+        '2001-01-01': '2001-01-01',
+        six.text_type('2016-01-10'): '2016-01-10',
+        None: None,
+        '': None,
+    }
+    field = serializers.DateField()
+
+
+class TestCustomInputFormatDateField(FieldValues):
+    """
+    Valid and invalid values for `DateField` with a custom input format.
+    """
+    valid_inputs = {
+        '1 Jan 2001': datetime.date(2001, 1, 1),
+    }
+    invalid_inputs = {
+        '2001-01-01': ['Date has wrong format. Use one of these formats instead: DD [Jan-Dec] YYYY.']
+    }
+    outputs = {}
+    field = serializers.DateField(input_formats=['%d %b %Y'])
+
+
+class TestCustomOutputFormatDateField(FieldValues):
+    """
+    Values for `DateField` with a custom output format.
+    """
+    valid_inputs = {}
+    invalid_inputs = {}
+    outputs = {
+        datetime.date(2001, 1, 1): '01 Jan 2001'
+    }
+    field = serializers.DateField(format='%d %b %Y')
+
+
+class TestNoOutputFormatDateField(FieldValues):
+    """
+    Values for `DateField` with no output format.
+    """
+    valid_inputs = {}
+    invalid_inputs = {}
+    outputs = {
+        datetime.date(2001, 1, 1): datetime.date(2001, 1, 1)
+    }
+    field = serializers.DateField(format=None)
