+++ tests/fields/test_time_field.py
@@ -1,0 +1,65 @@
+import datetime
+
+from rest_framework import serializers
+
+from .base import FieldValues
+
+
+class TestTimeField(FieldValues):
+    """
+    Valid and invalid values for `TimeField`.
+    """
+    valid_inputs = {
+        '13:00': datetime.time(13, 00),
+        datetime.time(13, 00): datetime.time(13, 00),
+    }
+    invalid_inputs = {
+        'abc': ['Time has wrong format. Use one of these formats instead: hh:mm[:ss[.uuuuuu]].'],
+        '99:99': ['Time has wrong format. Use one of these formats instead: hh:mm[:ss[.uuuuuu]].'],
+    }
+    outputs = {
+        datetime.time(13, 0): '13:00:00',
+        datetime.time(0, 0): '00:00:00',
+        '00:00:00': '00:00:00',
+        None: None,
+        '': None,
+    }
+    field = serializers.TimeField()
+
+
+class TestCustomInputFormatTimeField(FieldValues):
+    """
+    Valid and invalid values for `TimeField` with a custom input format.
+    """
+    valid_inputs = {
+        '1:00pm': datetime.time(13, 00),
+    }
+    invalid_inputs = {
+        '13:00': ['Time has wrong format. Use one of these formats instead: hh:mm[AM|PM].'],
+    }
+    outputs = {}
+    field = serializers.TimeField(input_formats=['%I:%M%p'])
+
+
+class TestCustomOutputFormatTimeField(FieldValues):
+    """
+    Values for `TimeField` with a custom output format.
+    """
+    valid_inputs = {}
+    invalid_inputs = {}
+    outputs = {
+        datetime.time(13, 00): '01:00PM'
+    }
+    field = serializers.TimeField(format='%I:%M%p')
+
+
+class TestNoOutputFormatTimeField(FieldValues):
+    """
+    Values for `TimeField` with a no output format.
+    """
+    valid_inputs = {}
+    invalid_inputs = {}
+    outputs = {
+        datetime.time(13, 00): datetime.time(13, 00)
+    }
+    field = serializers.TimeField(format=None)
